requirements.txt
aioredis==1.3.1
arrow==1.2.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
bidict==0.22.0
binaryornot==0.4.4
black==22.3.0
boto3==1.24.7
botocore==1.27.7
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.0.10
click==8.1.2
cloudinary==1.29.0
colorama==0.4.4
constantly==15.1.0
cookiecutter==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
dataclasses-json==0.5.7
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.3
dj-database-url==0.5.0
dj-rest-auth==2.2.3
Django==3.2.9
django-allauth==0.47.0
django-cors-headers==3.13.0
django-heroku==0.3.1
django-phonenumber-field==6.0.0
django-rest-auth==0.9.5
django-ses==3.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
facebook-sdk==3.1.0
filelock==3.4.0
google-api-core==2.10.1
google-api-python-client==2.63.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jinja2-time==0.2.0
jmespath==1.0.0
MarkupSafe==2.1.0
marshmallow==3.18.0
marshmallow-enum==1.5.1
msgpack==1.0.4
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.3
pathspec==0.9.0
pbr==5.7.0
pep8==1.7.1
phonenumbers==8.12.41
Pillow==9.2.0
platformdirs==2.4.0
protobuf==4.21.6
# psycopg2==2.9.3
# psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.5
python-engineio==4.3.4
python-facebook-api==0.15.0
python-slugify==6.1.2
python-socketio==5.7.1
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stevedore==3.5.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
twilio==7.8.0
txaio==22.2.1
typing-inspect==0.8.0
typing_extensions==4.3.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.18.2
virtualenv==20.10.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
whitenoise==5.3.0
wrapt==1.14.1
zope.interface==5.4.0
